# openapi.yaml
openapi: 3.0.0
info:
  x-skaffolder-id-project: 60794f54a3e0491123246c92
  title: Orus
  description: You can obtain JWT token from login API
  version: 0.1
servers:
  - url: http://localhost:3000/api
    description: Local server for testing
paths:
...
  "/taxpayerinfo/{id}":
    delete:
      x-skaffolder-id: 60794fb0a3e0491123246d10
      x-skaffolder-name: delete
      x-skaffolder-id-resource: 60794fa8a3e0491123246d07
      x-skaffolder-resource: Taxpayerinfo
      x-skaffolder-crudAction: delete
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION delete
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION delete
      tags:
        - Taxpayerinfo
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    get:
      x-skaffolder-id: 60794fb0a3e0491123246d0c
      x-skaffolder-name: get
      x-skaffolder-id-resource: 60794fa8a3e0491123246d07
      x-skaffolder-resource: Taxpayerinfo
      x-skaffolder-crudAction: get
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION get
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION get
      tags:
        - Taxpayerinfo
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id resource
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Taxpayerinfo"
    post:
      x-skaffolder-id: 60794fb0a3e0491123246d0b
      x-skaffolder-name: update
      x-skaffolder-id-resource: 60794fa8a3e0491123246d07
      x-skaffolder-resource: Taxpayerinfo
      x-skaffolder-crudAction: update
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION update
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION update
      tags:
        - Taxpayerinfo
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Taxpayerinfo"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Taxpayerinfo"
...
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Taxpayerinfo:
      x-skaffolder-id: 60794fa8a3e0491123246d07
      x-skaffolder-id-db: 60794f54a3e0491123246c93
      x-skaffolder-id-entity: 60794fa7a3e0491123246d04
      x-skaffolder-url: /taxpayerinfo
      properties:
        _id:
          type: string
          x-skaffolder-required: true
        BIRFormNumber:
          type: number
          x-skaffolder-id-attr: 60794fa7a3e0491123246d05
          x-skaffolder-type: Number
          x-skaffolder-required: true
        BirthDateOrgDate:
          type: integer
          x-skaffolder-id-attr: 60794fa7a3e0491123246d06
          x-skaffolder-type: Date
          x-skaffolder-required: true
      x-skaffolder-relations: 
      required:
        - _id
        - BIRFormNumber
        - BirthDateOrgDate
...